requirements.txt
bitdotio[psycopg2-binary]
pyaml
python-dotenv
pyserial